react packaging wip
