+++ packages/react/src/index.ts
@@ -1,2 +1,3 @@
-export { usePonderQuery, getQueryKey } from "./usePonderQuery.js";
 export { PonderProvider } from "./context.js";
+export { usePonderQuery } from "./hook.js";
+export { getQueryKey } from "./utils.js";
