+++ packages/react/src/utils.ts
@@ -1,0 +1,8 @@
+import { compileQuery } from "@ponder/client";
+
+export type SQLWrapper = Exclude<Parameters<typeof compileQuery>[0], string>;
+
+export function getQueryKey(query: SQLWrapper) {
+  const sql = compileQuery(query);
+  return [sql.sql, ...sql.params];
+}
